requirements.txt
# GrantAI Pro - Python Requirements
# Install with: pip install -r requirements.txt

# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
pymongo==4.5.0

# Authentication & Security
Werkzeug==2.3.7
python-dotenv==1.0.0

# Payment Processing
stripe==6.7.0

# HTTP Requests & External APIs
requests==2.31.0
urllib3==2.0.7

# Data Processing
python-dateutil==2.8.2
bson==0.5.10

# Production Server
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0

# Email Support (optional)
Flask-Mail==0.9.1

# API Documentation (optional)
flask-restx==1.2.0

# File Upload Support
Pillow==10.0.1

# Date/Time Utilities
pytz==2023.3

# JSON Web Tokens (if needed)
PyJWT==2.8.0

# Background Tasks (if needed)
celery==5.3.4
redis==5.0.1